Refactor comments and improve file handling
Implemented the 'list object' and the 'conversion of the program into a list' sections to create, fill a list and convert the program into it. Implemented also the relative functions to compile the program and make the parser of spaces and numbers. The exec function is a first basic implementation.